Move `VersionInfo`-related types to `cli/output-cache.ts`
This brings them out of the crowded all-purpose `util.ts` and into `cli/output-cache.ts` where they are exclusively used.
