+++ src/cli/output-cache.ts
@@ -3,9 +3,14 @@ import path from "path";
 
 import { getTemporaryDirectory } from "../actions-util";
 import { Env, getEnv } from "../environment";
-import { isPersistedVersionInfo } from "../util";
 
 import type { VersionInfo } from "./types";
+
+/** The persisted version together with the CLI path it was obtained from. */
+interface PersistedVersionInfo {
+  cmd: string;
+  version: VersionInfo;
+}
 
 /**
  * The name of the temporary file that backs the on-disk cache of
@@ -98,4 +103,36 @@ export function getCachedCodeQlVersion(
   // re-parse the environment variable.
   cachedCodeQlVersion = persisted.version;
   return cachedCodeQlVersion;
+}
+
+/**
+ * Determines whether a value is a `VersionInfo` object.
+ * @param x The value to test
+ */
+function isVersionInfo(x: unknown): x is VersionInfo {
+  const candidate = x as Partial<VersionInfo> | null;
+  return (
+    typeof candidate === "object" &&
+    candidate !== null &&
+    typeof candidate.version === "string" &&
+    (candidate.features === undefined ||
+      (typeof candidate.features === "object" &&
+        candidate.features !== null)) &&
+    (candidate.overlayVersion === undefined ||
+      typeof candidate.overlayVersion === "number")
+  );
+}
+
+/**
+ * Determines whether a value is a `PersistedVersionInfo` object.
+ * @param x The value to test
+ */
+function isPersistedVersionInfo(x: unknown): x is PersistedVersionInfo {
+  const candidate = x as Partial<PersistedVersionInfo> | null;
+  return (
+    typeof candidate === "object" &&
+    candidate !== null &&
+    typeof candidate.cmd === "string" &&
+    isVersionInfo(candidate.version)
+  );
 }
